Explicitly copy key data when composing prefix-compressed keys
Summary:
This fixes memory corruption where the data following the key was
overwritten.

Fix the priority comparison in ParallelIterator

This was requiring both the priority and the bytes to be minimal, when
really the priority is a tie-breaker when bytes are identical.

Properly handle empty iterators when setting up ParallelIterator

Test Plan: go test

Reviewers: tberman, jyeo

Reviewed By: tberman

CC: jimmy, justin

Differential Revision: http://phabricator.thefactory.com/D727