Yield all KVs in a block when searching for a lesser-than key
Summary:
This fixes T423, which would result in an empty iterator when
searching for an explicit key lesser than all items in a block.

Add similar checks to the other unit tests that weren't verifying
counts: an Assert in a foreach() will never be checked for empty
iterators!

Test Plan: unit tests

Reviewers: stephan, tberman

Reviewed By: tberman

CC: stephan, justin, tberman

Maniphest Tasks: T423

Differential Revision: http://phabricator.thefactory.com/D1354