Java implementation for datastore database
Summary: Full implementation of database with code that ensures memory tablets get pushed as file tablets when reaching a configurable size. This implementation also exposes an API to flush memory tablets and a Database.getOrElse() API.

Test Plan: Tested with included unit tests that use a combination of random data and the test data that is used for the C# unit tests. I plan to write a few more unit tests around thread safety and some volume tests with more random data that I plan to include with a subsequent patch.

Reviewers: peter

Reviewed By: peter

CC: serdar, justin, peter, tberman

Differential Revision: http://phabricator.thefactory.com/D1541