Performance improvements for MemoryTablet
Summary: Remove call to ConcurrentSkipListMap.size which turns out to be rather expensive. The side check is not needed as in the case of an empty memory tablet, the underlying iterator will correctly reflect an empty dataset. Also remove now unneeded empty iterator.

Test Plan: Unit tests and analytics benchmark

Reviewers: peter

Reviewed By: peter

CC: serdar, justin, tberman

Differential Revision: http://phabricator.thefactory.com/D1706